Fixed Tag Module and fixed welcome- and leave module

+++ src/modules/tags/events/tagsTextCommands.ts
@@ -1,7 +1,7 @@
-import { Events, Message } from "discord.js";
-import { convertToEmojiPng } from "../../../helper/emojis.js";;
-import { ExtendedClient } from "../../../types/client.js";
-import { database } from "../../../main/database.js";
+import {Events, Message} from "discord.js";
+import {convertToEmojiPng} from "../../../helper/emojis.js";
+import {ExtendedClient} from "../../../types/client.js";
+import {database} from "../../../main/database.js";
 
 export default {
     name: Events.MessageCreate,
@@ -13,51 +13,55 @@ export default {
      */
     async execute(message: Message, client: ExtendedClient) {
 
-        const data = await database.tags.findFirst({
-            where: {
-                GuildId: message.guild?.id,
-                TagId: message.content.split("!")[1]
+        try {
+            const data = await database.tags.findFirst({
+                where: {
+                    GuildId: message.guild?.id,
+                    TagId: message.content.split("!")[1]
+                }
+            });
+            if (message.author.id == client.user.id) return
+            if (!data) return;
+            if (data.IsEnabled == false) return;
+            if (data.PermissionRoleId) {
+                if (!client.user) throw new Error("Client user is not defined");
+
+                if (!message.member?.roles.cache.has(data.PermissionRoleId)) {
+                    const msg = message.reply({
+                        content: `## ${await convertToEmojiPng(
+                            "tag",
+                            client.user?.id
+                        )} You do not have the permission to use this tag.`
+                    });
+
+                    setTimeout(async () => {
+                        await message.delete();
+                        await (await msg).delete();
+                    }, 3000);
+                    return;
+                }
             }
-        });
 
-        if (!data) return;
-        if (data.IsEnabled == false) return;
-        if (data.PermissionRoleId) {
-            if (!client.user) throw new Error("Client user is not defined");
+            const messageData = await database.messageTemplates.findFirst({
+                where: {
+                    Name: data?.MessageTemplateId
+                }
+            });
 
-            if (!message.member?.roles.cache.has(data.PermissionRoleId)) {
-                const msg = message.reply({
-                    content: `## ${await convertToEmojiPng(
-                        "tag",
-                        client.user?.id
-                    )} You do not have the permission to use this tag.`
+            if (!messageData) return;
+
+            if (messageData.EmbedJSON) {
+                await message.reply({
+                    content: messageData.Content ?? "",
+                    embeds: [JSON.parse(messageData.EmbedJSON)]
                 });
-
-                setTimeout(async () => {
-                    await message.delete();
-                    await (await msg).delete();
-                }, 3000);
-                return;
+            } else {
+                await message.reply({
+                    content: messageData.Content ?? "",
+                });
             }
-        }
+        } catch (error) {
 
-        const messageData = await database.messageTemplates.findFirst({
-            where: {
-                Name: data.MessageId
-            }
-        });
-
-        if (!messageData) return;
-
-        if (messageData.EmbedJSON) {
-            message.reply({
-                content: messageData.Content ?? "",
-                embeds: [JSON.parse(messageData.EmbedJSON)]
-            });
-        } else {
-            message.reply({
-                content: messageData.Content ?? "",
-            });
         }
     }
 };
